[expo-notifications] Refactor
