+++ android/expoview/src/main/java/host/exp/exponent/services/ScopedExpoNotificationsService.java
@@ -4,7 +4,7 @@ import android.os.Parcel;
 
 import expo.modules.notifications.notifications.model.NotificationRequest;
 import expo.modules.notifications.notifications.service.ExpoNotificationsService;
-import host.exp.exponent.nextNotifications.ScopedNotificationRequest;
+import host.exp.exponent.notifications.ScopedNotificationRequest;
 
 public class ScopedExpoNotificationsService extends ExpoNotificationsService {
   @Override
