+++ android/expoview/src/main/java/versioned/host/exp/exponent/modules/universal/notifications/ScopedExpoNotificationPresentationModule.java
@@ -17,8 +17,8 @@ import expo.modules.notifications.notifications.model.NotificationRequest;
 import expo.modules.notifications.notifications.presentation.ExpoNotificationPresentationModule;
 import expo.modules.notifications.notifications.service.BaseNotificationsService;
 import host.exp.exponent.kernel.ExperienceId;
-import host.exp.exponent.nextNotifications.ScopedNotificationRequest;
-import host.exp.exponent.nextNotifications.ScopedNotificationsUtils;
+import host.exp.exponent.notifications.ScopedNotificationRequest;
+import host.exp.exponent.notifications.ScopedNotificationsUtils;
 
 public class ScopedExpoNotificationPresentationModule extends ExpoNotificationPresentationModule {
   private final ExperienceId mExperienceId;
